third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-mode-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-mode with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-mode">
<meta name="assert" content="mask-mode supports the full grammar '<masking-mode>#', where <masking-mode> is 'alpha | luminance | match-source'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("mask-mode", "alpha");
test_valid_value("mask-mode", "luminance");
test_valid_value("mask-mode", "match-source");

test_valid_value("mask-mode", "alpha, alpha");
test_valid_value("mask-mode", "match-source, luminance");
</script>
</body>
</html>
